+++ js/data.js
@@ -1,0 +1,77 @@
+
+const avatars = [
+  'img/avatars/user01.png',
+  'img/avatars/user02.png',
+  'img/avatars/user03.png',
+  'img/avatars/user04.png',
+  'img/avatars/user05.png',
+  'img/avatars/user06.png',
+  'img/avatars/user07.png',
+  'img/avatars/user08.png',
+  'img/avatars/user09.png',
+  'img/avatars/user10.png',
+];
+
+const titles = [
+  'Дом для отпуска на берегу озера Маслозеро',
+  'Отель Карелия',
+  'Гостевой дом Онежский берег',
+  'Эко Отель Большая Медведица',
+  'Гостиница Петрозаводск',
+  'Марциальные ключи',
+  'Гостиница Северная',
+  'Общежитие Карелреспотребсоюза',
+  'Отель Онежский замок',
+  'Дом для хороших гостей',
+];
+
+const types = [
+  'palace',
+  'flat',
+  'house',
+  'bungalow',
+  'hotel',
+];
+
+const checkins = [
+  '12:00',
+  '13:00',
+  '14:00',
+];
+
+const checkouts = [
+  '12:00',
+  '13:00',
+  '14:00',
+];
+
+
+const descriptions = [
+  'Семейный дом — Дуплекс',
+  'Дом с 2 спальнями',
+  'Стандартный номер-студио',
+  'Люкс',
+  'Лофт',
+  'Классический четырехместный номер',
+  'Номер-студио с видом на озеро',
+  'Двухуровневые апартаменты',
+  'Дом с 1 спальней',
+  'Дом для отпуска',
+];
+
+const photos = [
+  'https://assets.htmlacademy.ru/content/intensive/javascript-1/keksobooking/duonguyen-8LrGtIxxa4w.jpg',
+  'https://assets.htmlacademy.ru/content/intensive/javascript-1/keksobooking/brandon-hoogenboom-SNxQGWxZQi0.jpg',
+  'https://assets.htmlacademy.ru/content/intensive/javascript-1/keksobooking/claire-rendall-b6kAwr1i0Iw.jpg',
+];
+
+const featuresList = [
+  'wifi',
+  'dishwasher',
+  'parking',
+  'washer',
+  'elevator',
+  'conditioner',
+];
+
+export {avatars, titles, types, checkins, checkouts, descriptions, photos, featuresList};
